requirements.txt
Flask~=2.0.2
requests~=2.26.0
mysql-connector-python~=8.0.27